+++ mji-craftwork/src/solver/bruce_force.rs
@@ -67,7 +67,7 @@ where
         }
 
         self.data.foreach(|r| {
-            if r.craft_time == 0 || remain < r.craft_time as u16 {
+            if r.craft_time == 0 || remain < r.craft_time {
                 return;
             }
             if !limit.check(r) || last.id == r.id {
@@ -86,7 +86,7 @@ where
             let s = self.data.state(id, demands[id]);
             let demand_sub = current.get_produce(s.id()) * info.workers;
             let val = simulate(&info, &s, demand_sub);
-            let current = current.push(r.id, val, r.cost, r.craft_time as u16);
+            let current = current.push(r.id, val, r.cost, r.craft_time);
 
             if current.get_time() > limit.time - 4 {
                 // 当前工序结束
